requirements.txt
discord-py
asyncio
datetime
Cybernator
six
pyowm
pyshorteners
mctools